add feature profile scrap

+++ app/Http/Controllers/Blogs/BlogController.php
@@ -5,12 +5,17 @@ use App\Http\Controllers\Controller;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 use App\Models\Blog;
+
+use League\CommonMark\CommonMarkConverter;
 
 
 class BlogController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Request $request)
     {
         $blogs = Blog::query();
@@ -30,11 +35,10 @@ class BlogController extends Controller
         $blogs = $blogs->with('owner')->latest()->paginate(10)->withQueryString();
 
         return view('blogs.index', [
-            'title' => 'Filtered Blogs',
+            'title' => 'Blogs',
             'blogs' => $blogs,
         ]);
     }
-
 
     public function create()
     {
@@ -59,19 +63,43 @@ class BlogController extends Controller
 
     public function show(Blog $blog)
     {
-        $blog->load('owner', 'comments.replies');
+        // Load comments, top-level only, with their replies and commenters
+        $blog->load([
+            'comments' => function ($query) {
+            $query->whereNull('parent_id')->latest()
+                ->with(['commenter', 'replies.commenter']);
+            },
+            'owner'
+        ]);
+
+        // Load current user's reaction if logged in
+        if (auth()->check()) {
+            $userHandle = auth()->user()->user_handle;
+            $blog->user_reaction = $blog->reactions()->where('user_id', $userHandle)->first();
+        } else {
+            $blog->user_reaction = null;
+        }
+
+        // Convert Markdown to HTML
+        $converter = new CommonMarkConverter([
+            'html_input' => 'escape',
+            'allow_unsafe_links' => false,
+        ]);
+    
+        $blog->parsed_content = $converter->convertToHtml($blog->content);
+    
         return view('blogs.show', compact('blog'));
     }
 
     public function edit(Blog $blog)
     {
-        //$this->authorize('update', $blog);
+        $this->authorize('update', $blog);
         return view('blogs.edit', compact('blog'));
     }
 
     public function update(Request $request, Blog $blog)
     {
-        //$this->authorize('update', $blog);
+        $this->authorize('update', $blog);
 
         $validated = $request->validate([
             'title' => 'required|unique:blogs,title,' . $blog->id,
@@ -85,7 +113,7 @@ class BlogController extends Controller
 
     public function destroy(Blog $blog)
     {
-        //$this->authorize('delete', $blog);
+        $this->authorize('delete', $blog);
         $blog->delete();
         return redirect()->route('blogs.index')->with('success', 'Blog deleted!');
     }
